add: gameModel & separate view files
