+++ pong_project/pong/views/__init__.py
@@ -1,0 +1,4 @@
+from .auth_views import login_api, logout_api, signup_api
+from .user_views import user_info_api, update_user_info_api, user_count
+from .index_view import index
+from .health_check_view import health_check
