+++ pong_project/pong/views/health_check_view.py
@@ -1,0 +1,4 @@
+from django.http import JsonResponse
+
+def health_check(request):
+		return JsonResponse({'status': 'healthy'})
