+++ pong_project/pong/views/index_view.py
@@ -1,0 +1,4 @@
+from django.shortcuts import render
+
+def index(request):
+	return render(request, 'pong/index.html')
